% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check for NULL.}

\item{y}{Default value to return if x is NULL.}
}
\value{
x if not NULL, otherwise y.
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the
right-hand side. Similar to the \verb{??} operator in other languages.
}
\examples{
NULL \%||\% "default"
# Returns "default"

"value" \%||\% "default"
# Returns "value"

}
\keyword{internal}
